Convert common memory to uint8 to match BF spec

+++ src/synesthesiapkg/common.nim
@@ -2,7 +2,7 @@ import strformat
 
 type
   BFCore* = ref object
-   memory*: array[1024, int]
+   memory*: array[1024, uint8]
    pc*, ap*: int
 
   BFTokenKind* = enum
